Refactor tests int sub-packages
